pyproject.toml
[project]
name = "whisperlivekit"
version = "0.2.0"
description = "Test faster-whisper Hindi transcription with a browser mic interface"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "faster-whisper>=1.1.0",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
